+++ src/index.js
@@ -19,7 +19,8 @@ export { DateInput } from './components/DateInput'
 export { MoneyInput } from './components/MoneyInput'
 export { PhoneInput } from './components/PhoneInput'
 export { TimeInput } from './components/TimeInput'
-export { CountrySelect } from './components/CountrySelect'
+export { CountryInput } from './components/CountryInput'
+export { CountryInput as CountrySelect } from './components/CountryInput'
 export { ContentTransition } from './components/ContentTransition'
 export { SkeletonPulse, SkeletonLine } from './components/Skeleton'
 export {
